Clean-ups and renames
