+++ osu.Game/Screens/Ranking/ResultModeTabControl.cs
@@ -3,12 +3,11 @@
 
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.UserInterface;
-using osu.Game.Screens.Ranking.Types;
 using osuTK;
 
 namespace osu.Game.Screens.Ranking
 {
-    public class ResultModeTabControl : TabControl<IResultType>
+    public class ResultModeTabControl : TabControl<IResultPageInfo>
     {
         public ResultModeTabControl()
         {
@@ -20,9 +19,9 @@ namespace osu.Game.Screens.Ranking
             TabContainer.Padding = new MarginPadding(5);
         }
 
-        protected override Dropdown<IResultType> CreateDropdown() => null;
+        protected override Dropdown<IResultPageInfo> CreateDropdown() => null;
 
-        protected override TabItem<IResultType> CreateTabItem(IResultType value) => new ResultModeButton(value)
+        protected override TabItem<IResultPageInfo> CreateTabItem(IResultPageInfo value) => new ResultModeButton(value)
         {
             Anchor = TabContainer.Anchor,
             Origin = TabContainer.Origin
