+++ osu.Game/Screens/Ranking/Results.cs
@@ -19,7 +19,6 @@ using osu.Game.Graphics.UserInterface;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Scoring;
-using osu.Game.Screens.Ranking.Types;
 
 namespace osu.Game.Screens.Ranking
 {
@@ -281,6 +280,6 @@ namespace osu.Game.Screens.Ranking
             }, true);
         }
 
-        protected abstract IEnumerable<IResultType> CreateResultTypes();
+        protected abstract IEnumerable<IResultPageInfo> CreateResultTypes();
     }
 }
