+++ osu.Game/Screens/Ranking/ResultsPage.cs
@@ -12,9 +12,9 @@ using osu.Game.Scoring;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Screens.Ranking.Pages
+namespace osu.Game.Screens.Ranking
 {
-    public class ResultsPage : Container
+    public abstract class ResultsPage : Container
     {
         protected readonly ScoreInfo Score;
         protected readonly WorkingBeatmap Beatmap;
@@ -23,7 +23,7 @@ namespace osu.Game.Screens.Ranking.Pages
 
         protected override Container<Drawable> Content => content;
 
-        public ResultsPage(ScoreInfo score, WorkingBeatmap beatmap)
+        protected ResultsPage(ScoreInfo score, WorkingBeatmap beatmap)
         {
             Score = score;
             Beatmap = beatmap;
